refactor: errors in different files
