+++ src/errors/400/data-not-object.error.ts
@@ -1,0 +1,7 @@
+import { BadRequestError } from './index.error'
+
+export class DataNotObjectError extends BadRequestError {
+  constructor() {
+    super(`Given data is not JavaScript object type`) // i.e) User email is not present
+  }
+}
