+++ src/errors/400/index.error.ts
@@ -1,0 +1,14 @@
+import { BadRequestException } from '@nestjs/common'
+
+/**
+ * @class BadRequestError
+ * @description
+ * Thrown when certain data is missing when it is supposed to be present.
+ * @example
+ * throw new BadRequestError(`User email is not present`)
+ */
+export class BadRequestError extends BadRequestException {
+  constructor(message: string) {
+    super(message)
+  }
+}
